Module obj::obj

Parses .obj format which stores 3D mesh data

Structs

Group
A group which contains multiple range of points, lines and polygons

Obj
Low-level Rust binding for .obj format.

Range
A struct which represent [start, end) range.

Enums

Line
The Line type.

Polygon
The Polygon type.

Functions

load_obj
Parses a wavefront .obj format

Type Definitions

Point
The Point type which stores the index of the position vector.