Enum obj::obj::Polygon

pub enum Polygon {
    P(Vec<u32>),
    PT(Vec<(u32, u32)>),
    PN(Vec<(u32, u32)>),
    PTN(Vec<(u32, u32, u32)>),
}

The Polygon type.

Variants

P

A polygon which contains only the position data of each vertex.

PT

A polygon which contains both position and texture coordinate data of each vertex.

PN

A polygon which contains both position and normal data of each vertex.

PTN

A polygon which contains all position, texture coordinate and normal data of each vertex.