Module obj::raw::object

Parses .obj format which stores 3D mesh data

Structs

Group
A group which contains ranges of points, lines and polygons

Range
A struct which represent [start, end) range.

RawObj
Low-level Rust binding for .obj format.

Enums

Line
The Line type.

Polygon
The Polygon type.

Functions

parse_obj
Parses a wavefront .obj format.

Type Definitions

Point
The Point type which stores the index of the position vector.