Enum obj::raw::object::Polygon

pub enum Polygon {
    P(Vec<usize>),
    PT(Vec<(usize, usize)>),
    PN(Vec<(usize, usize)>),
    PTN(Vec<(usize, usize, usize)>),
}

The Polygon type.

Variants

P

A polygon which contains only the position data of each vertex.

PT

A polygon which contains both position and texture coordinate data of each vertex.

PN

A polygon which contains both position and normal data of each vertex.

PTN

A polygon which contains all position, texture coordinate and normal data of each vertex.